feat(main): implement standard cat
Implement the standard execution of cat, which simply prints the contents of given files as they
appear. If "-" is provided as an argument, print standard input line by line (i.e., lines are
printed as they are entered in standard input and broken with a newline character). If no arguments
at all are given, run as if only a single "-" was passed as an argument.